hardware/python_code/python_code/requirements.txt
dnspython==2.3.0
dotenv==0.0.5
numpy==1.24.3
pymongo==4.3.3
RPi.GPIO==0.7.1
